refactor(modeler): remove old app/ folder
