+++ apps/miranum-modeler/src/adapter/in/index.ts
@@ -1,4 +1,4 @@
 export * from "./commands";
-export * from "./webview";
 export * from "./custom_text_editor";
+export * from "./webview";
 export * from "./workspace";
