--- apps/miranum-modeler/src/app/lib/index.ts
@@ -1,3 +1,0 @@
-export * from "./Watcher";
-export * from "./StandardTextEditor";
-export * from "./vscode-settings";
