+++ apps/miranum-modeler/src/application/useCases/index.ts
@@ -1,4 +1,5 @@
-export * from "./restoreWebview";
-export * from "./sendToWebview";
+export * from "./displayEditor";
 export * from "./miranumConfig";
 export * from "./queries";
+export * from "./restoreEditor";
+export * from "./syncDocument";
